refactor: DRY up fix
