+++ cmd/influxd/upgrade/logging_windows.go
@@ -8,39 +8,23 @@ import (
 	"go.uber.org/zap"
 )
 
+// Work around a bug in zap's handling of absolute paths on Windows.
+// See https://github.com/uber-go/zap/issues/621
+
+const FakeWindowsScheme = "winfile"
+
 func init() {
-	// NOTE: Work around a bug in zap's handling of absolute paths on Windows.
-	// See https://github.com/uber-go/zap/issues/621
 	newWinFileSink := func(u *url.URL) (zap.Sink, error) {
 		// Remove leading slash left by url.Parse()
 		return os.OpenFile(u.Path[1:], os.O_WRONLY|os.O_APPEND|os.O_CREATE, 0644)
 	}
-	zap.RegisterSink("winfile", newWinFileSink)
+	zap.RegisterSink(FakeWindowsScheme, newWinFileSink)
 }
 
-func buildLogger(options *logOptions, verbose bool) (*zap.Logger, error) {
-	config := zap.NewProductionConfig()
-
-	config.Level = zap.NewAtomicLevelAt(options.logLevel)
-	if verbose {
-		config.Level.SetLevel(zap.DebugLevel)
-	}
-
-	logPath, err := filepath.Abs(options.logPath)
+func (o *logOptions) zapSafeLogPath() (string, error) {
+	logPath, err := filepath.Abs(o.logPath)
 	if err != nil {
-		return nil, err
+		return "", err
 	}
-	logPath = "winfile:///" + logPath
-
-	config.OutputPaths = append(config.OutputPaths, logPath)
-	config.ErrorOutputPaths = append(config.ErrorOutputPaths, logPath)
-
-	log, err := config.Build()
-	if err != nil {
-		return nil, err
-	}
-	if verbose {
-		log.Warn("--verbose is deprecated, use --log-level=debug instead")
-	}
-	return log, nil
+	return FakeWindowsScheme + ":///" + logPath, nil
 }
